pyproject.toml
[project]
name = "rtx-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "chromadb>=1.3.5",
    "docling>=2.31.0",
    "httpx>=0.28.0",
    "markdownify>=0.14.1",
    "fastapi>=0.122.0",
    "langchain-experimental>=0.4.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "langchain>=1.1.0",
    "langfuse>=3.10.1",
    "langgraph>=1.0.4",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "mcp[cli]>=1.23.1",
    "neo4j~=5.28.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.3.5",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pypdf>=5.0.0",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "rapidfuzz>=3.0.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sse-starlette>=3.0.3",
    "uvicorn>=0.38.0",
    "crawl4ai>=0.7.8",
    "pypdf2>=3.0.1",
]
